\name{AIME-package}
\alias{AIME-package}
\alias{AIME}
\docType{package}
\title{
Autoencoder-based Integrative Multi-omics data Embedding
}
\description{
A nonlinear data embedding method using modified auto encoder
}
\details{
Autoencoder-based Integrative Multi-omics data Embedding (AIME), to extract data representation for omics data integrative analysis. The method can achieve informative data embedding, rank features in terms of their contributions, and find pairs of features from the two data types that are related to each other through the data embedding. 
}
\author{
Tianwei Yu <yutianwei@cuhk.edu.cn>

Maintainer: Tianwei Yu <yutianwei@cuhk.edu.cn>
}
\keyword{ package }

